Switched argument type again

+++ src/main/java/net/niccossystem/perfectarena/arena/PvPArena.java
@@ -64,7 +64,7 @@ public class PvPArena {
      */
     public void removePlayer(Player player) {
         players.remove(player);
-        player.damage(Integer.MAX_VALUE);
+        player.damage(Double.MAX_VALUE);
     }
 
     public Location getRandomSpawnPoint() {
